gpu engine gets correct collections/almost correct outputs

+++ cpp/tradeup_engine/src/items/market_item_memory_flat_data.hpp
@@ -11,7 +11,7 @@ struct MarketItemMemoryFlatData {
     std::vector<int> outcomeCollectionsStartIndices;
     std::vector<int> outcomeCollectionsEndIndices;
     
-    std::vector<uint64_t> outputItemIds;
+    std::vector<int> outputItemIds;
     std::vector<int> outputItemIdsStartIndices;
     std::vector<int> outputItemIdsEndIndices;
